+++ RecoEgamma/EgammaElectronProducers/plugins/alpaka/ElectronNHitSeedAlpakaProducer.cc
@@ -41,7 +41,7 @@
 
 #include "PixelMatchingAlgo.h"
 
-using Vector3d = Eigen::Matrix<double, 3, 1>;
+using Vector3f = Eigen::Matrix<float, 3, 1>;
 
 namespace ALPAKA_ACCELERATOR_NAMESPACE {
 
@@ -57,7 +57,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
     void produce(edm::StreamID sid, device::Event& event, device::EventSetup const& iSetup) const override {
       auto vprim_ = event.get(beamSpotToken_).position();
       GlobalPoint vprim(vprim_.x(), vprim_.y(), vprim_.z());
-      Vector3d vertex{vprim.x(), vprim.y(), vprim.z()};
+      Vector3f vertex{static_cast<float>(vprim.x()), static_cast<float>(vprim.y()), static_cast<float>(vprim.z())};
 
       const std::vector<reco::SuperClusterRef>& superClusterRefVec = event.get(superClustersTokens_);
       int32_t superClusterCollectionSize = superClusterRefVec.size();
@@ -101,9 +101,9 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
         const auto& rot0 = recHit0.det()->surface().rotation().z();
         viewSeeds[i].hit0detectorID() = (recHit0.geographicalId().subdetId() == PixelSubdetector::PixelBarrel) ? 1 : 0;
         viewSeeds[i].hit0isValid() = recHit0.isValid();
-        viewSeeds[i].hit0Pos() = Eigen::Vector3d(pos0.x(), pos0.y(), pos0.z());
-        viewSeeds[i].surf0Pos() = Eigen::Vector3d(surf0.x(), surf0.y(), surf0.z());
-        viewSeeds[i].surf0Rot() = Eigen::Vector3d(rot0.x(), rot0.y(), rot0.z());
+        viewSeeds[i].hit0Pos() = Eigen::Vector3f(pos0.x(), pos0.y(), pos0.z());
+        viewSeeds[i].surf0Pos() = Eigen::Vector3f(surf0.x(), surf0.y(), surf0.z());
+        viewSeeds[i].surf0Rot() = Eigen::Vector3f(rot0.x(), rot0.y(), rot0.z());
 
         // Hit 1
         ++hitIt;
@@ -113,9 +113,9 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
         const auto& rot1 = recHit1.det()->surface().rotation().z();
         viewSeeds[i].hit1detectorID() = (recHit1.geographicalId().subdetId() == PixelSubdetector::PixelBarrel) ? 1 : 0;
         viewSeeds[i].hit1isValid() = recHit1.isValid();
-        viewSeeds[i].hit1Pos() = Eigen::Vector3d(pos1.x(), pos1.y(), pos1.z());
-        viewSeeds[i].surf1Pos() = Eigen::Vector3d(surf1.x(), surf1.y(), surf1.z());
-        viewSeeds[i].surf1Rot() = Eigen::Vector3d(rot1.x(), rot1.y(), rot1.z());
+        viewSeeds[i].hit1Pos() = Eigen::Vector3f(pos1.x(), pos1.y(), pos1.z());
+        viewSeeds[i].surf1Pos() = Eigen::Vector3f(surf1.x(), surf1.y(), surf1.z());
+        viewSeeds[i].surf1Rot() = Eigen::Vector3f(rot1.x(), rot1.y(), rot1.z());
 
         // Hit 2
         if (initialSeedRef.nHits() > 2) {
@@ -127,9 +127,9 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
           viewSeeds[i].hit2detectorID() =
               (recHit2.geographicalId().subdetId() == PixelSubdetector::PixelBarrel) ? 1 : 0;
           viewSeeds[i].hit2isValid() = recHit2.isValid();
-          viewSeeds[i].hit2Pos() = Eigen::Vector3d(pos2.x(), pos2.y(), pos2.z());
-          viewSeeds[i].surf2Pos() = Eigen::Vector3d(surf2.x(), surf2.y(), surf2.z());
-          viewSeeds[i].surf2Rot() = Eigen::Vector3d(rot2.x(), rot2.y(), rot2.z());
+          viewSeeds[i].hit2Pos() = Eigen::Vector3f(pos2.x(), pos2.y(), pos2.z());
+          viewSeeds[i].surf2Pos() = Eigen::Vector3f(surf2.x(), surf2.y(), surf2.z());
+          viewSeeds[i].surf2Rot() = Eigen::Vector3f(rot2.x(), rot2.y(), rot2.z());
         } else {
           // Zero initialization
           viewSeeds[i].hit2Pos().setZero();
